+++ CHANGELOG.md
@@ -7,5 +7,6 @@
 - Administracyjna lista zwierząt (/admin/pets) + przyciski do narzędzi
 - Administracyjna lista typów zwierząt (/admin/pet-types)
 - Administracyjna lista typów aktywności (/admin/action-types)
+- Administracyjna lista użytkowników (/admin/users)
 - Menu administracyjne